+++ e2e-etl-python/files/terraform-data.tf
@@ -1,0 +1,14 @@
+locals {
+  terraform-data = {
+    id             = local.id
+    name           = var.name
+    tags           = local.tags
+    current_region = data.aws_region.current.name
+  }
+}
+
+resource "local_file" "terraform-data" {
+  filename = "${path.module}/.terraform-data.json"
+  content  = jsonencode(local.terraform-data)
+}
+